Apply OrbDancer empowerment Offensive Buffs

+++ VBusiness/Weapons/CommonWeapons/BaseTemporaryBuff.cs
@@ -9,6 +9,6 @@ namespace VBusiness.Weapons
 		public abstract int Duratation { get; }
 		public abstract int Cooldown { get; }
 		public abstract IDisposable ApplyTemporaryBuff(VLoadout loadout);
-		public double AbilityUptime => Math.Min(Duratation / Cooldown, 1);
+		public double AbilityUptime => Math.Min(((double)Duratation) / Cooldown, 1);
 	}
 }
